♻️ Refactor :  와이어프레임대로 nav 수정
