+++ src/app/(dashboard)/analysis/recommendation/page.tsx
@@ -1,5 +1,15 @@
-function Page() {
-  return <>recommendation</>;
-}
+import ReviewCountChart from './_blocks/ReviewCountChart';
+import ReviewCountChartFilter from './_blocks/ReviewCountChartFilter';
 
-export default Page;
+export default async function ReviewPage() {
+  return (
+    <>
+      <section className="flex w-full">
+        <ReviewCountChartFilter />
+      </section>
+      <section className="flex h-[500px] w-full">
+        <ReviewCountChart />
+      </section>
+    </>
+  );
+}
