--- src/app/(dashboard)/product/page.tsx
@@ -1,3 +1,0 @@
-export default function ProductPage() {
-  return <>Product Page</>;
-}
