--- src/app/(dashboard)/store/page.tsx
@@ -1,3 +1,0 @@
-export default function StorePage() {
-  return <>Store Page</>;
-}
