+++ src/global/NavigationDrawer/constants.ts
@@ -1,3 +1,3 @@
 import PATH from '@/src/shared/constants/path';
 
-export const navigations = [PATH.activate, PATH.user, PATH.product, PATH.store, PATH.review];
+export const navigations = [PATH.activate, PATH.user, PATH.review];
